pyproject.toml
[project]
name = "rideshare"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mysql-connector-python>=9.2.0",
]
